organizing vuex using modules
