+++ src/assets/global.css
@@ -115,6 +115,7 @@ textarea {
   font-size: 100%;
   line-height: 1.15;
   margin: 0;
+  border-radius: 4px;
 }
 button,
 input {
@@ -205,7 +206,7 @@ select {
     no-repeat right 12px center;
   background-size: 8px 10px;
   border: solid 1px rgba(0, 0, 0, 0.4);
-  border-radius: 0;
+  border-radius: 4px;
   -webkit-appearance: none;
   -moz-appearance: none;
   appearance: none;
