+++ src/store/index.js
@@ -1,9 +1,9 @@
 import { createStore } from 'vuex'
-import EventService from '@/services/EventService'
+import * as user from '@/store/modules/user'
+import * as event from '@/store/modules/event'
 
 export default createStore({
   state: {
-    user: { id: '1', name: 'Dan Ray Rollan'},
     categories: [
       'consistency',
       'nature',
@@ -13,63 +13,9 @@ export default createStore({
       'community',
       'open-source'
     ],
-    events: [],
-    eventsTotal: 0,
-    event: {}
-  },
-  getters: {
-    catLength: state => {
-      return state.categories.length
-    },
-    getEventById: state => id => {
-      return state.events.find(event => event.id === id)
-    }
-  },
-  mutations: {
-    ADD_EVENT(state, event) {
-      state.events.push(event)
-    },
-    SET_EVENTS(state, events) {
-      state.events = events
-    }, 
-    SET_EVENTS_TOTAL(state, eventsTotal) { 
-      state.eventsTotal = eventsTotal
-    },
-    SET_EVENT(state, event) {
-      state.event = event
-    }
-  },
-  actions: {
-    createEvent({ commit }, event) {
-      return EventService.postEvent(event).then(()=> {
-        commit('ADD_EVENT', event) 
-      })
-    },
-    fetchEvents({ commit }, { perPage, page }) {
-      EventService.getEvents(perPage, page)
-        .then(response => {
-          commit('SET_EVENTS_TOTAL', parseInt(response.headers['x-total-count']))
-          commit('SET_EVENTS', response.data)
-        })
-        .catch(error => {
-          console.log("An e rror occurred: " + error.response)
-        })
-    },
-    fetchEvent({commit, getters}, id) {
-      var event = getters.getEventById(id)
-      if (event) {
-        commit('SET_EVENT', event)
-      } else {
-      EventService.getEvent(id)
-        .then(response => {
-          commit('SET_EVENT', response.data)
-        })
-        .catch(error => {
-          console.log('There was an error:', error.response)
-        })
-      }
-    }
   },
   modules: {
+    user,
+    event
   }
 })
