+++ src/store/modules/user.js
@@ -1,0 +1,3 @@
+export const state = {
+  user: { id: '1', name: 'Dan Ray Rollan'},
+}
